Added basic scrolling to listbox. Changed scroll events to occur when a widget is focused, not when the mouse is inside.

+++ M_ApplicationClass.pde
@@ -89,7 +89,7 @@ class ApplicationClass {
       m_currentScreen.onMouseClick();
   }
   
-  void onMouseWheel(float wheelCount){
+  void onMouseWheel(int wheelCount){
     if (m_currentScreen != null)
       m_currentScreen.getOnMouseWheelEvent().raise(new MouseWheelEventInfoType(mouseX, mouseY, wheelCount, m_currentScreen));
